fix: restructure mutations into separate files
